experiments/requirements.txt
anthropic>=0.40.0
langfuse>=2.0.0
python-dotenv>=1.0.0
